+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/order/request/UpdatedOrder.kt
@@ -176,7 +176,7 @@ data class UpdatedOrder(
 		val giftCard: Boolean? = null,
 
 		val selectedOptions: List<OrderItemSelectedOption>? = null,
-		val combinationId: Int? = null,
+		val combinationId: Long? = null,
 		val taxes: List<OrderItemTax>? = null,
 		val dimensions: ProductDimensions? = null,
 		val discountsAllowed: Boolean? = null,
