+++ src/test/kotlin/com/ecwid/apiclient/v3/entity/CartsTest.kt
@@ -506,7 +506,7 @@ class CartsTest : BaseEntityTest() {
 				generateDateSelectedOption(),
 				generateFilesSelectedOption()
 			),
-			combinationId = randomId(),
+			combinationId = randomId().toLong(),
 			taxes = listOf(
 				generateTestOrderItemTax()
 			),
