+++ src/test/kotlin/com/ecwid/apiclient/v3/util/OrderUtils.kt
@@ -253,7 +253,7 @@ private fun generateTestOrderItem() = UpdatedOrder.OrderItem(
 		generateDateSelectedOption(),
 		generateFilesSelectedOption()
 	),
-	combinationId = randomId(),
+	combinationId = randomId().toLong(),
 	taxes = listOf(
 		generateTestOrderItemTax(),
 		generateTestOrderItemTax()
